app1/app1_client/requirements.txt
opencv-contrib-python
numpy
flask
requests
Pillow